+++ docs/source/conf.py
@@ -6,24 +6,25 @@
 
 import os
 import sys
-sys.path.insert(0, os.path.abspath(os.path.join('..','..')))
+
+sys.path.insert(0, os.path.abspath(os.path.join("..", "..")))
 
 
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = 'demops'
-copyright = '2022, Patrik Spiess'
-author = 'Patrik Spiess'
+project = "demops"
+copyright = "2022, Patrik Spiess"
+author = "Patrik Spiess"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = ['sphinx.ext.autodoc']
+extensions = ["sphinx.ext.autodoc"]
 exclude_patterns = []
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = 'sphinx_rtd_theme'
-html_static_path = ['_static']
+html_theme = "sphinx_rtd_theme"
+html_static_path = ["_static"]